Grant invitation workflow

+++ src/main/java/org/codealpha/gmsservice/validators/DashboardValidator.java
@@ -53,7 +53,7 @@ public class DashboardValidator {
     }
 
     private void _validateUser(Long userId, User user) {
-        if (userId != user.getId()) {
+        if (userId.longValue() != user.getId().longValue()) {
             throw new ResourceNotFoundException("Invalid authorization");
         }
     }
